scripts/quote-scraper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
gutenbergpy>=0.3.5
anthropic>=0.18.0
python-dotenv>=1.0.0
tqdm>=4.66.0
rapidfuzz>=3.0.0
pytest>=7.4.0
